feat: implement custom hook to manage localStorage
