+++ homeworks/pavel-didenko/20-react-api/src/App.jsx
@@ -1,10 +1,11 @@
 import './App.css';
 import Task from './components/Task/Task';
 import TaskCreator from './components/TaskCreator/TaskCreator';
-import {useEffect, useReducer, useCallback} from 'react';
+import { useReducer, useCallback} from 'react';
 import reducer from './functions/reducer';
 import { MainContext } from './contexts/mainContext';
 import ClearCompletedTasksButton from './components/ClearCompletedTasksButton/ClearCompletedTasksButton';
+import useSaveToLocalStorage from './hooks/useSaveToLocalStorage';
 
 const App = function () {
   const [state, dispatch] = useReducer(reducer, []);
@@ -26,17 +27,7 @@ const App = function () {
     });
   }, [state]);
 
-  useEffect(() => {
-    const storedTasks = localStorage.getItem('tasks');
-    if (storedTasks) {
-      const tasksObject = JSON.parse(storedTasks);
-      dispatch({type: 'tasks_extracted', tasks: tasksObject});
-    }
-  }, []);
-
-  useEffect(() => {
-    localStorage.setItem('tasks', JSON.stringify(state));
-  }, [state]);
+  useSaveToLocalStorage(state, dispatch);
 
   function createTask(newTaskText) {
     dispatch({type: 'task_added', text: newTaskText});
