+++ homeworks/pavel-didenko/20-react-api/src/hooks/useSaveToLocalStorage.js
@@ -1,0 +1,16 @@
+import { useEffect } from "react";
+
+export default function useSaveToLocalStorage(state, dispatch) {
+
+  useEffect(() => {
+    const storedTasks = localStorage.getItem('tasks');
+    if (storedTasks) {
+      const tasksObject = JSON.parse(storedTasks);
+      dispatch({type: 'tasks_extracted', tasks: tasksObject});
+    }
+  }, []);
+
+  useEffect(() => {
+    localStorage.setItem('tasks', JSON.stringify(state));
+  }, [state]);
+}
